api/omdatabase.cc,api/sortable-serialise.cc,backends/chert/, backends/flint/,backends/remote/remote-database.cc, common/serialise-double.cc,common/str.cc, matcher/exactphrasepostlist.cc,net/remoteconnection.cc, net/remoteserver.cc: Replace C-style casts.  Add a few missing casts. Tweak types used in a few places.
git-svn-id: file:///home/xapian-svn/xapian/trunk@12768 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/backends/flint/flint_spelling.cc
@@ -541,7 +541,7 @@ FlintSpellingTermList::next()
 	    throw Xapian::DatabaseCorruptError("Bad spelling termlist");
 	current_term.resize(byte(data[p++]) ^ MAGIC_XOR_VALUE);
     }
-    size_t add;
+    unsigned add;
     if (p == data.size() ||
 	(add = byte(data[p]) ^ MAGIC_XOR_VALUE) >= data.size() - p)
 	throw Xapian::DatabaseCorruptError("Bad spelling termlist");
